fixed various variable types

+++ 2DGameEngine/src/Components/ProjectileComponent.h
@@ -13,7 +13,7 @@ struct ProjectileComponent
 	}
 
 	int hitDamage;
-	unsigned projectileDurationS;
+	float projectileDurationS;
 	bool isFriendly;
 	float lifeTime;
 };
